Feature/gestion usuarios basicos

+++ ed-lib-common/src/main/java/com/eduplanner/ed_lib_common/dto/RegisterStaffDTO.java
@@ -35,7 +35,7 @@ public class RegisterStaffDTO {
 
     private String disabilities;
 
-    private Byte stratum;
+    private Integer stratum;
 
     private String populationType;
